+++ HeishaMon/src/common/strnicmp.cpp
