+++ HeishaMon/src/common/strnicmp.h
